+++ tests/contrib/test_pydantic.py
@@ -7,6 +7,7 @@ from tortoise.contrib.pydantic import pydantic_model_creator, pydantic_queryset_
 
 class TestPydantic(test.TestCase):
     async def asyncSetUp(self) -> None:
+        await super(TestPydantic, self).asyncSetUp()
         self.Event_Pydantic = pydantic_model_creator(Event)
         self.Event_Pydantic_List = pydantic_queryset_creator(Event)
         self.Tournament_Pydantic = pydantic_model_creator(Tournament)
@@ -1036,6 +1037,7 @@ class TestPydantic(test.TestCase):
 
 class TestPydanticCycle(test.TestCase):
     async def asyncSetUp(self) -> None:
+        await super(TestPydanticCycle, self).asyncSetUp()
         self.Employee_Pydantic = pydantic_model_creator(Employee)
 
         self.root = await Employee.create(name="Root")